+++ src/game.c
@@ -86,11 +86,12 @@ void game_init(game_t *p_game) {
 	text_renderer_init(&p_game->text_renderer, p_game->renderer, &p_game->font);
 
 	game_load_asset(p_game, "./res/blocks.bmp",    ASSET_BLOCKS);
-	game_load_asset(p_game, "./res/icons.bmp",     ASSET_ICONS);
+	game_load_asset(p_game, "./res/gold.bmp",      ASSET_GOLD);
 	game_load_asset(p_game, "./res/particles.bmp", ASSET_PARTICLES);
 	game_load_asset(p_game, "./res/arrows.bmp",    ASSET_ARROWS);
 	game_load_asset(p_game, "./res/outline.bmp",   ASSET_OUTLINE);
 	game_load_asset(p_game, "./res/menu.bmp",      ASSET_MENU);
+	game_load_asset(p_game, "./res/icons.bmp",     ASSET_ICONS);
 
 	block_renderer_init(&p_game->block_renderer, p_game->renderer, &p_game->assets[ASSET_BLOCKS]);
 	world_init(&p_game->world, &p_game->block_renderer, CAM_POS_X, CAM_POS_Y);
@@ -98,7 +99,32 @@ void game_init(game_t *p_game) {
 	p_game->img.outline = image_new(&p_game->assets[ASSET_OUTLINE],
 	                             CENTERX(p_game->assets[ASSET_OUTLINE].rect.w), 2);
 
-	p_game->img.gold = image_new(&p_game->assets[ASSET_ICONS], 2, 2);
+	p_game->img.gold = image_new(&p_game->assets[ASSET_GOLD], 2, 2);
+
+	p_game->img.icons[1] = image_new(&p_game->assets[ASSET_ICONS],
+	                                 CENTERX(p_game->assets[ASSET_ICONS].rect.w / 3),
+	                                 SCREEN_H - p_game->assets[ASSET_ICONS].rect.h / 2 - 2);
+	p_game->img.icons[1].dest.w /= 3;
+	p_game->img.icons[1].dest.h /= 2;
+	p_game->img.icons[1].src.w  /= 3;
+	p_game->img.icons[1].src.h  /= 2;
+	p_game->img.icons[1].src.x   = p_game->assets[ASSET_ICONS].rect.w / 3;
+
+	p_game->img.icons[0] = image_new(&p_game->assets[ASSET_ICONS],
+	                                 p_game->img.icons[1].dest.x - 24,
+	                                 p_game->img.icons[1].dest.y);
+	p_game->img.icons[0].dest.w = p_game->img.icons[1].dest.w;
+	p_game->img.icons[0].dest.h = p_game->img.icons[1].dest.h;
+	p_game->img.icons[0].src    = p_game->img.icons[1].src;
+	p_game->img.icons[0].src.x  = 0;
+
+	p_game->img.icons[2] = image_new(&p_game->assets[ASSET_ICONS],
+	                                 p_game->img.icons[1].dest.x + 24,
+	                                 p_game->img.icons[1].dest.y);
+	p_game->img.icons[2].dest.w = p_game->img.icons[1].dest.w;
+	p_game->img.icons[2].dest.h = p_game->img.icons[1].dest.h;
+	p_game->img.icons[2].src    = p_game->img.icons[1].src;
+	p_game->img.icons[2].src.x  = p_game->assets[ASSET_ICONS].rect.w / 3 * 2;
 
 	int x = p_game->img.outline.dest.x - p_game->assets[ASSET_ARROWS].rect.w / 2 - 1;
 	int y = p_game->img.outline.dest.y + CENTER(p_game->assets[ASSET_ARROWS].rect.h / 2,
@@ -166,8 +192,9 @@ void game_free_assets(game_t *p_game) {
 const char *game_mode_name(game_t *p_game) {
 	switch (p_game->mode) {
 	case MODE_VIEWING:  return "Viewing";
-	case MODE_PLACING:  return "Placing";
 	case MODE_DELETING: return "Deleting";
+	case MODE_PLACING:  return "Placing";
+	case MODE_CHANGING: return "Changing";
 
 	default: assert(0 && "game_mode_name(): non-existant game mode");
 	}
@@ -247,6 +274,21 @@ void game_render_ui(game_t *p_game) {
 
 	button_render(&p_game->button.menu, p_game->renderer);
 
+	for (size_t i = 0; i < SIZE_OF(p_game->img.icons); ++ i) {
+		bool highlight_this = p_game->mode == i + 1;
+		if (highlight_this)
+			p_game->img.icons[i].src.y = p_game->assets[ASSET_ICONS].rect.h / 2;
+		else
+			image_set_alpha(&p_game->img.icons[i], 128);
+
+		image_render(&p_game->img.icons[i], p_game->renderer);
+
+		if (highlight_this)
+			p_game->img.icons[i].src.y = 0;
+		else
+			image_set_alpha(&p_game->img.icons[i], SDL_ALPHA_OPAQUE);
+	}
+
 	if (p_game->mode == MODE_PLACING)
 		game_render_placing_ui(p_game);
 
@@ -263,8 +305,17 @@ void game_render_placing_ui(game_t *p_game) {
 	if (p_game->cursor_block.cost > p_game->gold)
 		block_renderer_set_alpha(&p_game->block_renderer, 128);
 
-	block_renderer_render_plain_sprite(&p_game->block_renderer, p_game->img.outline.dest.x + 2,
-	                                   p_game->img.outline.dest.y + 2,
+	int sel_x = CENTERX(BLOCK_SIZE);
+	int sel_y = p_game->img.outline.dest.y + 2;
+
+	int offset = PREV_NEXT_BLOCKS_OFFSET / SHOP_REPOS_TIME * p_game->shop_repos_timer;
+
+	if (p_game->shop_repos_dir == DIR_LEFT)
+		sel_x -= offset;
+	else
+		sel_x += offset;
+
+	block_renderer_render_plain_sprite(&p_game->block_renderer, sel_x, sel_y,
 	                                   block_get_sprite(&p_game->cursor_block), BLOCK_SIZE);
 
 	if (p_game->cursor_block.cost > p_game->gold)
@@ -274,6 +325,47 @@ void game_render_placing_ui(game_t *p_game) {
 	snprintf(gold_str, sizeof(gold_str), "%zu", block_type_cost(blocks_shop[p_game->shop_pos]));
 
 	game_render_text_center_x(p_game, ENDY(p_game->img.outline.dest) + 1, gold_str);
+
+	block_t prev = p_game->cursor_block, next = p_game->cursor_block;
+
+	if (p_game->shop_pos + 1 < SIZE_OF(blocks_shop)) {
+		block_set_type(&next, blocks_shop[p_game->shop_pos + 1]);
+
+		int x = CENTERX(BLOCK_SIZE) + PREV_NEXT_BLOCKS_OFFSET;
+		int y = p_game->img.outline.dest.y + 2;
+
+		if (p_game->shop_repos_dir == DIR_LEFT)
+			x -= offset;
+		else
+			x += offset;
+
+		block_renderer_set_alpha(&p_game->block_renderer, 128);
+
+		block_renderer_render_plain_sprite(&p_game->block_renderer, x, y,
+		                                   block_get_sprite(&next), BLOCK_SIZE);
+
+		block_renderer_set_alpha(&p_game->block_renderer, SDL_ALPHA_OPAQUE);
+	}
+
+	if (p_game->shop_pos > 0) {
+		block_set_type(&prev, blocks_shop[p_game->shop_pos - 1]);
+
+		int x = CENTERX(BLOCK_SIZE) - PREV_NEXT_BLOCKS_OFFSET;
+		int y = p_game->img.outline.dest.y + 2;
+
+		if (p_game->shop_repos_dir == DIR_LEFT)
+			x -= offset;
+		else
+			x += offset;
+
+		block_renderer_set_alpha(&p_game->block_renderer, 128);
+
+		block_renderer_render_plain_sprite(&p_game->block_renderer, x, y,
+		                                   block_get_sprite(&prev), BLOCK_SIZE);
+
+		block_renderer_set_alpha(&p_game->block_renderer, SDL_ALPHA_OPAQUE);
+	}
+
 }
 
 void game_render_gold_ui(game_t *p_game) {
@@ -288,7 +380,8 @@ void game_render_gold_ui(game_t *p_game) {
 }
 
 void game_render_mode_ui(game_t *p_game) {
-	game_render_text_center_x(p_game, SCREEN_H - p_game->font.ch_h * 2, game_mode_name(p_game));
+	game_render_text_center_x(p_game, SCREEN_H - p_game->font.ch_h * 2 -
+	                          p_game->assets[ASSET_ICONS].rect.h / 2, game_mode_name(p_game));
 }
 
 void game_render_paused_ui(game_t *p_game) {
@@ -364,20 +457,32 @@ void game_events_ingame(game_t *p_game) {
 		button_events(&p_game->button.right, &p_game->event, &p_game->mouse);
 
 		if (button_pressed(&p_game->button.left)) {
+			if (p_game->shop_repos_timer > 0)
+				return;
+
 			if (p_game->shop_pos == 0)
 				p_game->shop_pos = SIZE_OF(blocks_shop) - 1;
 			else
 				-- p_game->shop_pos;
 
-			p_game->cursor_block.type = blocks_shop[p_game->shop_pos];
+			block_set_type(&p_game->cursor_block, blocks_shop[p_game->shop_pos]);
+
+			p_game->shop_repos_timer = SHOP_REPOS_TIME;
+			p_game->shop_repos_dir   = DIR_LEFT;
 
 			return;
 		} else if (button_pressed(&p_game->button.right)) {
+			if (p_game->shop_repos_timer > 0)
+				return;
+
 			++ p_game->shop_pos;
 			if (p_game->shop_pos >= SIZE_OF(blocks_shop))
 				p_game->shop_pos = 0;
 
-			p_game->cursor_block.type = blocks_shop[p_game->shop_pos];
+			block_set_type(&p_game->cursor_block, blocks_shop[p_game->shop_pos]);
+
+			p_game->shop_repos_timer = SHOP_REPOS_TIME;
+			p_game->shop_repos_dir   = DIR_RIGHT;
 
 			return;
 		}
@@ -390,6 +495,25 @@ void game_events_ingame(game_t *p_game) {
 	}
 
 	switch (p_game->event.type) {
+	case SDL_MOUSEWHEEL:
+		if (p_game->event.wheel.y > 0) {
+			switch (p_game->mode) {
+			case MODE_VIEWING:  p_game->mode = MODE_CHANGING; break;
+			case MODE_CHANGING: p_game->mode = MODE_DELETING; break;
+
+			default: ++ p_game->mode;
+			}
+		} else if (p_game->event.wheel.y < 0) {
+			switch (p_game->mode) {
+			case MODE_VIEWING:  p_game->mode = MODE_DELETING; break;
+			case MODE_DELETING: p_game->mode = MODE_CHANGING; break;
+
+			default: -- p_game->mode;
+			}
+		}
+
+		break;
+
 	case SDL_KEYDOWN:
 		switch (p_game->event.key.keysym.sym) {
 		case SDLK_RETURN:
@@ -400,9 +524,10 @@ void game_events_ingame(game_t *p_game) {
 
 			break;
 
-		case SDLK_ESCAPE:    p_game->mode = MODE_VIEWING;  break;
-		case SDLK_e:         p_game->mode = MODE_PLACING;  break;
-		case SDLK_BACKSPACE: p_game->mode = MODE_DELETING; break;
+		case SDLK_ESCAPE: p_game->mode = MODE_VIEWING;  break;
+		case SDLK_1:      p_game->mode = MODE_DELETING; break;
+		case SDLK_2:      p_game->mode = MODE_PLACING;  break;
+		case SDLK_3:      p_game->mode = MODE_CHANGING; break;
 
 		case SDLK_r:
 			if (p_game->mode == MODE_PLACING) {
@@ -444,6 +569,9 @@ void game_events_paused(game_t *p_game) {
 
 void game_update(game_t *p_game) {
 	++ p_game->tick;
+
+	if (p_game->shop_repos_timer > 0)
+		-- p_game->shop_repos_timer;
 
 	if (p_game->mode == MODE_PLACING)
 		p_game->can_place = world_cursor_tile(&p_game->world)->top.type == BLOCK_AIR;