+++ src/image.c
@@ -20,6 +20,10 @@ image_t image_new(asset_t *p_asset, int p_x, int p_y) {
 	};
 }
 
+void image_set_alpha(image_t *p_image, Uint8 p_a) {
+	SDL_SetTextureAlphaMod(p_image->texture, p_a);
+}
+
 void image_render(image_t *p_image, SDL_Renderer *p_renderer) {
 	SDL_RenderCopy(p_renderer, p_image->texture, &p_image->src, &p_image->dest);
 }